requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Talisman==1.1.0
Flask-Compress==1.14
gunicorn==21.2.0
Werkzeug==3.0.1
SQLAlchemy==2.0.27
python-magic==0.4.27
qrcode==7.4.2
Pillow==10.2.0